@@ -1,58 +1,32 @@
 source 'https://rubygems.org'
 
 
-# Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '~> 6.1', '>= 6.0.3.2'
-# Dependence required by rails
-gem 'actionpack', '~> 6.1', '>= 6.0.3.2'
-# Use SCSS for stylesheets
+gem 'rails', '~> 7.2', '>= 7.2.2.1'
+gem 'actionpack', '~> 7.2'
 gem 'sass-rails', '~> 6.0'
-# Use CoffeeScript for .coffee assets and views
 gem 'coffee-rails', '~> 5.0.0'
-# See https://github.com/rails/execjs#readme for more supported runtimes
-# gem 'therubyracer', platforms: :ruby
-
-# Use jquery as the JavaScript library
-gem 'jquery-rails'
-# Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder', '~> 2.0'
-# bundle exec rake doc:rails generates the API under doc/api.
-gem 'sdoc', '~> 2.0', '>= 2.0.3'
-gem 'devise', '~> 4.7', '>= 4.7.1'
-gem 'simple_form', '>= 3.5.0'
-gem 'httparty', '~> 0.18.0' #HTTP Requests
-gem 'carrierwave', '~> 2.1' #Uploader
-gem 'mini_magick', '~> 4.9', '>= 4.9.5' #Manipulates images
-gem 'carrierwave-bombshelter', '~> 0.2.2' #Protects from bomb images
-gem "recaptcha", require: "recaptcha/rails" #reCaptcha
-
-gem 'alertifyjs-rails'
-
-# Use ActiveModel has_secure_password
-# gem 'bcrypt', '~> 3.1.7'
-
-# Use Unicorn as the app server
-# gem 'unicorn'
-
-# Use Capistrano for deployment
-# gem 'capistrano-rails', group: :development
+gem 'jquery-rails', '~> 4.6', '>= 4.6.0'
+gem 'jbuilder', '~> 2.13'
+gem 'sdoc', '~> 2.6', '>= 2.6.1'
+gem 'devise', '~> 4.9', '>= 4.9.4'
+gem 'simple_form', '~> 5.3', '>= 5.3.1'
+gem 'httparty', '~> 0.22.0'
+gem 'carrierwave', '~> 3.1', '>= 3.1.1'
+gem 'mini_magick', '~> 5.1', '>= 5.1.2'
+gem 'carrierwave-bombshelter', '~> 0.2.2' 
+gem 'recaptcha', '~> 5.19', require: "recaptcha/rails" 
+gem 'alertifyjs-rails', '~> 1.12', '>= 1.11.2'
+gem 'pg', '~> 1.5', '>= 1.5.9'
+gem 'puma', '~> 6.6'
+gem 'nokogiri', '~> 1.18', '>= 1.18.3'
 
 group :development, :test do
-  # Call 'byebug' anywhere in the code to stop execution and get a debugger console
-  gem 'byebug'
-  gem 'rspec-rails', '~> 3.9'
+  gem 'rspec-rails', '~> 7.1', '>= 7.1.1'
   gem 'factory_girl_rails', '~> 4.7.0'
-  gem 'capybara', '~> 2.7.1'
+  gem 'capybara', '~> 3.40'
 end
 
 group :development do
-  # Access an IRB console on exception pages or by using <%= console %> in views
-  gem 'web-console', '~> 4.0'
-
-  # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
-  gem 'spring', '~> 2.0.2'
+  gem 'web-console', '~> 4.2', '>= 4.2.1'
+  gem 'spring', '~> 4.2.1'
 end
-
-
-# Added at 2017-09-21 13:51:12 -0500 by valentina:
-gem "pg", "~> 1.2.3"
